database restructured
table users modified
