+++ dependencies.py
@@ -1,0 +1,7 @@
+from sqlalchemy import Column,Date,Integer, ForeignKey, String
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import relationship
+from sqlalchemy import create_engine
+from psycopg2 import connect
+import sys
+from psycopg2.extensions import ISOLATION_LEVEL_AUTOCOMMIT
